requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
django-admin-ordering==0.12.0
django-ckeditor==5.9.0
django-content-editor==2.0.0
django-imagefield==0.11.3
django-js-asset==1.2.2
django-tree-queries==0.3.3
et-xmlfile==1.0.1
feincms3==0.33.0
feincms3-meta==3.0.1
feincms3-sites==0.10.0
form-designer==0.16.1
geographiclib==1.50
geopy==1.21.0
html-sanitizer==1.9.0
idna==2.8
jdcal==1.4.1
lxml==4.5.0
openpyxl==3.0.3
Pillow==7.0.0
pytz==2019.3
requests==2.22.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.8
xlsxdocument==1.3.5
